+++ raiden/tests/integration/fixtures/transport.py
@@ -1,7 +1,7 @@
 import pytest
 
 from raiden.constants import DISCOVERY_DEFAULT_ROOM, PATH_FINDING_BROADCASTING_ROOM
-from raiden.network.transport import MatrixNode as MatrixTransportNode
+from transport.matrix.transport import MatrixNode as MatrixTransportNode
 from raiden.tests.fixtures.variables import TransportProtocol
 from raiden.tests.utils.transport import generate_synapse_config, matrix_server_starter
