+++ raiden/tests/integration/test_recovery.py
@@ -6,7 +6,6 @@ from raiden import waiting
 from raiden.api.python import RaidenAPI
 from raiden.app import App
 from raiden.message_handler import MessageHandler
-from raiden.network.transport import UDPTransport
 from raiden.raiden_event_handler import RaidenEventHandler
 from raiden.tests.utils.events import search_for_item
 from raiden.tests.utils.network import CHAIN
@@ -22,6 +21,7 @@ from raiden.transfer.state_change import (
     ContractReceiveChannelSettled,
 )
 from raiden.utils import create_default_identifier
+from transport.udp.udp_transport import UDPTransport
 
 
 @pytest.mark.parametrize("deposit", [10])
