+++ raiden/tests/unit/test_udp_transport.py
@@ -6,7 +6,7 @@ from gevent import server
 from raiden.constants import UINT64_MAX
 from raiden.messages import SecretRequest
 from raiden.network.throttle import TokenBucket
-from transport.udp import UDPTransport
+from transport.udp.udp_transport import UDPTransport
 from raiden.tests.utils.factories import ADDR, UNIT_SECRETHASH, make_address
 from raiden.tests.utils.mocks import MockRaidenService
 from raiden.tests.utils.transport import MockDiscovery
