+++ raiden/ui/app.py
@@ -1,3 +1,4 @@
+import json
 import os
 import sys
 from typing import Any, Callable, Dict, TextIO
@@ -6,14 +7,13 @@ from urllib.parse import urlparse
 import click
 import filelock
 import structlog
-from eth_utils import to_canonical_address, to_normalized_address
-from web3 import HTTPProvider, Web3
-from definitions import ROOT_DIR
-import json
 from eth_utils import encode_hex
+from eth_utils import to_canonical_address, to_normalized_address
+from raiden_contracts.constants import ID_TO_NETWORKNAME
+from raiden_contracts.contract_manager import ContractManager
+from web3 import HTTPProvider, Web3
 
-from transport.matrix import MatrixLayer as MatrixTransportLayer
-
+from definitions import ROOT_DIR
 from raiden.accounts import AccountManager
 from raiden.constants import (
     RAIDEN_DB_VERSION,
@@ -49,8 +49,7 @@ from raiden.ui.startup import (
 )
 from raiden.utils import BlockNumber, pex, split_endpoint
 from raiden.utils.typing import Address, Optional, PrivateKey, Tuple
-from raiden_contracts.constants import ID_TO_NETWORKNAME
-from raiden_contracts.contract_manager import ContractManager
+from transport.matrix.layer import MatrixLayer as MatrixTransportLayer
 
 log = structlog.get_logger(__name__)
