--- transport/udp/__init__.py
@@ -1,1 +1,0 @@
-from transport.udp import UDPTransport  # NOQA
